feat: move filter APIs to filters module
