--- apps/web/core/blocks-sdk/index.ts
@@ -1,1 +1,0 @@
-export * as TableBlockSdk from './table';
