+++ apps/web/core/blocks/data/to-query-string.test.ts
@@ -1,11 +1,11 @@
 import { SYSTEM_IDS } from '@geogenesis/sdk';
 import { describe, expect, it } from 'vitest';
 
-import { FilterString, createFiltersFromFilterString, createGraphQLStringFromFilters } from './table';
+import { queryStringFromFilters } from './to-query-string';
 
-describe('TableBlock SDK', () => {
+describe('to-query-string', () => {
   it('Builds a graphql query from table block filters', () => {
-    const stringFilter = createGraphQLStringFromFilters([
+    const stringFilter = queryStringFromFilters([
       {
         columnId: 'type',
         value: 'Value 1',
@@ -17,7 +17,7 @@ describe('TableBlock SDK', () => {
       `triples: { some: { attributeId: { equalTo: "type" }, textValue: { equalToInsensitive: "Value 1"} } }`
     );
 
-    const entityFilter = createGraphQLStringFromFilters([
+    const entityFilter = queryStringFromFilters([
       {
         columnId: 'type',
         value: 'id 1',
@@ -28,7 +28,7 @@ describe('TableBlock SDK', () => {
     // Don't support relation filters in the query yet
     expect(entityFilter).toEqual(`and: []`);
 
-    const nameFilter = createGraphQLStringFromFilters([
+    const nameFilter = queryStringFromFilters([
       {
         columnId: SYSTEM_IDS.NAME_ATTRIBUTE,
         value: 'id 1',
@@ -38,7 +38,7 @@ describe('TableBlock SDK', () => {
 
     expect(nameFilter).toEqual(`name: { startsWithInsensitive: "id 1" }`);
 
-    const andFilter = createGraphQLStringFromFilters([
+    const andFilter = queryStringFromFilters([
       {
         columnId: SYSTEM_IDS.TYPES_ATTRIBUTE,
         value: 'Value 1',
@@ -55,7 +55,7 @@ describe('TableBlock SDK', () => {
       `and: [{ versionTypes: { some: { type: { entityId: {equalTo: "Value 1" } } } } }, { name: { startsWithInsensitive: "id 1" } }]`
     );
 
-    const spaceFilter = createGraphQLStringFromFilters([
+    const spaceFilter = queryStringFromFilters([
       {
         columnId: SYSTEM_IDS.SPACE_FILTER,
         valueType: 'TEXT',
@@ -65,69 +65,19 @@ describe('TableBlock SDK', () => {
 
     expect(spaceFilter).toEqual(
       `versionSpaces: {
-          some: {
-            spaceId: { equalTo: "0x0000000000000000000000000000000000000000" }
-          }
-        }`
+        some: {
+          spaceId: { equalTo: "0x0000000000000000000000000000000000000000" }
+        }
+      }`
     );
 
-    const nullTypeIdFilter = createGraphQLStringFromFilters([]);
+    const nullTypeIdFilter = queryStringFromFilters([]);
 
     expect(nullTypeIdFilter).toEqual('');
   });
 
-  it('Builds the TableBlockStore filters data structure from the filter string', async () => {
-    /**
-     * There are several combinations of filters that can be applied to a table block.
-     * 1. String field with a string value
-     * 2. Entity field with an entity ID as the value
-     * 3. String field targeting the Name column
-     * 4. A null type id
-     *
-     * These four combinations can also be used together with an "and" filter
-     */
-    const filter: FilterString = {
-      where: {
-        spaces: ['0x0000000000000000000000000000000000000000'],
-        AND: [
-          {
-            attribute: SYSTEM_IDS.TYPES_ATTRIBUTE,
-            is: SYSTEM_IDS.SCHEMA_TYPE,
-          },
-          {
-            attribute: SYSTEM_IDS.NAME_ATTRIBUTE,
-            is: 'name',
-          },
-        ],
-      },
-    };
-
-    const stringFilter = await createFiltersFromFilterString(JSON.stringify(filter));
-
-    expect(stringFilter).toEqual([
-      {
-        columnId: SYSTEM_IDS.SPACE_FILTER,
-        value: '0x0000000000000000000000000000000000000000',
-        valueType: 'RELATION',
-        valueName: null,
-      },
-      {
-        columnId: SYSTEM_IDS.TYPES_ATTRIBUTE,
-        value: SYSTEM_IDS.SCHEMA_TYPE,
-        valueType: 'RELATION',
-        valueName: 'Type',
-      },
-      {
-        columnId: SYSTEM_IDS.NAME_ATTRIBUTE,
-        value: 'name',
-        valueType: 'TEXT',
-        valueName: null,
-      },
-    ]);
-  });
-
   it('Builds a graphql query from table block filters for the postgraphile-based substreams API', () => {
-    const stringFilter = createGraphQLStringFromFilters([
+    const stringFilter = queryStringFromFilters([
       {
         columnId: 'type',
         value: 'Value 1',
@@ -137,7 +87,7 @@ describe('TableBlock SDK', () => {
 
     expect(stringFilter).toMatchSnapshot();
 
-    const entityFilter = createGraphQLStringFromFilters([
+    const entityFilter = queryStringFromFilters([
       {
         columnId: 'type',
         value: 'id 1',
@@ -147,7 +97,7 @@ describe('TableBlock SDK', () => {
 
     expect(entityFilter).toMatchSnapshot();
 
-    const nameFilter = createGraphQLStringFromFilters([
+    const nameFilter = queryStringFromFilters([
       {
         columnId: SYSTEM_IDS.NAME_ATTRIBUTE,
         value: 'id 1',
@@ -157,7 +107,7 @@ describe('TableBlock SDK', () => {
 
     expect(nameFilter).toMatchSnapshot();
 
-    const andFilter = createGraphQLStringFromFilters([
+    const andFilter = queryStringFromFilters([
       {
         columnId: 'type',
         value: 'Value 1',
@@ -177,7 +127,7 @@ describe('TableBlock SDK', () => {
 
     expect(andFilter).toMatchSnapshot();
 
-    const spaceFilter = createGraphQLStringFromFilters([
+    const spaceFilter = queryStringFromFilters([
       {
         columnId: SYSTEM_IDS.SPACE_FILTER,
         valueType: 'TEXT',
@@ -187,7 +137,7 @@ describe('TableBlock SDK', () => {
 
     expect(spaceFilter).toMatchSnapshot();
 
-    const nullTypeIdFilter = createGraphQLStringFromFilters([]);
+    const nullTypeIdFilter = queryStringFromFilters([]);
 
     expect(nullTypeIdFilter).toEqual('');
   });
